Adds settings menu functionality
Implements the settings menu, which includes the ability to display and interact with settings.

This commit introduces new UI prefabs for category and module items, and refactors the menu scripts to load and display settings. It also creates new functions in the settings manager to differentiate between core and non-core settings.